refactor: add models dir
