+++ src/database/data-source.ts
@@ -5,8 +5,8 @@ import { Channel } from '../channels/models/channel.entity';
 import { Image } from '../images/models/image.entity';
 import { Invite } from '../invites/invite.entity';
 import { Message } from '../messages/message.entity';
-import { ProposalAction } from '../proposals/proposal-action.entity';
-import { Proposal } from '../proposals/proposal.entity';
+import { ProposalAction } from '../proposals/models/proposal-action.entity';
+import { Proposal } from '../proposals/models/proposal.entity';
 import { Permission } from '../roles/models/permission.entity';
 import { Role } from '../roles/models/role.entity';
 import { User } from '../users/user.entity';
