+++ src/images/models/image.entity.ts
@@ -7,7 +7,7 @@ import {
   UpdateDateColumn,
 } from 'typeorm';
 import { Message } from '../../messages/message.entity';
-import { Proposal } from '../../proposals/proposal.entity';
+import { Proposal } from '../../proposals/models/proposal.entity';
 
 @Entity()
 export class Image {
