+++ src/users/user.entity.ts
@@ -10,7 +10,7 @@ import {
 import { ChannelMember } from '../channels/models/channel-member.entity';
 import { Invite } from '../invites/invite.entity';
 import { Message } from '../messages/message.entity';
-import { Proposal } from '../proposals/proposal.entity';
+import { Proposal } from '../proposals/models/proposal.entity';
 import { Role } from '../roles/models/role.entity';
 import { Vote } from '../votes/vote.entity';
